+++ app/Http/Controllers/UserController.php
@@ -4,18 +4,28 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\UserRequest;
 use App\Models\User;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
+
+    use AuthorizesRequests; // Permite verificar los permisos del usuario autenticado
+
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
+
+        $this->authorize('user-list'); // Verifica que solo los usuarios con el permiso user-list puedan acceder a esta ruta
+
         $texto = $request->input('texto');
 
-        $registros = User::where('name', 'LIKE', "%{$texto}%")
+        $registros = User::with('roles')
+            ->where('name', 'LIKE', "%{$texto}%")
             ->orWhere('email', 'LIKE', "%{$texto}%")
             ->orderBy('id', 'DESC')
             ->paginate(10);
@@ -28,7 +38,11 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('usuario.action');
+        $this->authorize('user-create');
+
+        $roles = Role::all();
+
+        return view('usuario.action', compact('roles'));
     }
 
     /**
@@ -36,12 +50,17 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        $this->authorize('user-create');
+
         $registro = new User();
+
         $registro->name = $request->input('name');
         $registro->email = $request->input('email');
-        $registro->password = bcrypt($request->input('password'));
+        $registro->password = Hash::make($request->input('password'));
         $registro->activo= $request->input('activo');
         $registro->save();
+
+        $registro->assignRole($request->input('role')); // Asigna el rol seleccionado en el formulario 
 
         return redirect()->route('usuarios.index')->with('mensaje', 'Registro ' . $registro->name . ' creado correctamente.');
     }
@@ -59,11 +78,15 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('user-edit');
+
+        $roles = Role::all();
+
         $registro = User::findOrFail($id);
         // con User $usuario como parametros hubiera sido: 
         // $registro = User::findOrFail($usuario->id);
 
-        return view('usuario.action', compact('registro'));
+        return view('usuario.action', compact('registro', 'roles'));
     }
 
     /**
@@ -71,13 +94,22 @@ class UserController extends Controller
      */
     public function update(UserRequest $request, $id)
     {
+        $this->authorize('user-edit');
+
         $registro = User::findOrFail($id);
 
         $registro->name = $request->input('name');
         $registro->email = $request->input('email');
-        $registro->password = bcrypt($request->input('password'));
+
+        // Si existe contenido en el input de password, se actualiza, sino, se deja el mismo que ya tiene
+        if ($request->filled('password')) {
+            $registro->password = Hash::make($request->input('password'));
+        }
+
         $registro->activo = $request->input('activo');
         $registro->save();
+
+        $registro->syncRoles($request->input('role')); // Actualiza el rol del usuario en base a lo seleccionado en el formulario, si tenia antes otro rol, lo elimina y asigna el nuevo
 
         return redirect()->route('usuarios.index')->with('mensaje', 'Registro ' . $registro->name . ' actualizado correctamente.');
     }
@@ -87,6 +119,8 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('user-delete');
+
         $registro = User::findOrFail($id);
         $registro->delete();
 
@@ -96,6 +130,9 @@ class UserController extends Controller
     // Funcionalidad para hacer un soft delete
     // Se podria igualmente haber trabajado con $id como en los metodos anteriores
     public function toggleStatus(User $usuario) {
+
+        $this->authorize('user-activate');
+
         $usuario->activo = !$usuario->activo; // Cambia el estado activo del usuario al valor contrario
         $usuario->save();
